Feature/DaysCouting

+++ src/service/user.service.ts
@@ -35,7 +35,7 @@ const getUserByEmail = async (email: string): Promise<User> => {
 
   return user;
 };
-const getUserById = async (userId: number) => {
+const getUserById = async (userId: number): Promise<User> => {
   const user = await prisma.user.findUnique({
     where: { id: userId, isDeleted: false },
   });
@@ -44,19 +44,19 @@ const getUserById = async (userId: number) => {
 const updateUserById = async (
   userId: number,
   updateData: Prisma.UserUpdateInput,
-) => {
+): Promise<User> => {
   return await prisma.user.update({
     where: { id: userId, isDeleted: false },
     data: updateData,
   });
 };
-const restoreUserById = async (userId: number) => {
+const restoreUserById = async (userId: number): Promise<User> => {
   return await prisma.user.update({
     where: { id: userId, isDeleted: true },
     data: { isDeleted: false, deleteAt: null },
   });
 };
-const softDeleteUserById = async (userId: number) => {
+const softDeleteUserById = async (userId: number): Promise<User> => {
   const user = await getUserById(userId);
   if (!user) throw new ApiError(httpStatus.NOT_FOUND, "User not found");
   return await prisma.user.update({
@@ -66,18 +66,23 @@ const softDeleteUserById = async (userId: number) => {
       deleteAt: dayjs().toDate(),
     },
   });
-
-  user;
 };
 
-const hardDeleteUserById = async (userId: number) => {
-  const user = await getUserById(userId);
+const hardDeleteUserById = async (userId: number): Promise<User> => {
+  const user = await prisma.user.findUnique({ where: { id: userId } });
   if (!user) throw new ApiError(httpStatus.NOT_FOUND, "User not found");
-  await prisma.user.delete({
-    where: { id: userId },
-  });
+  return await prisma.user.delete({ where: { id: userId } });
+};
 
-  return user;
+const cleanUpDeletedUsers = async (beforeDate: Date) => {
+  return await prisma.user.deleteMany({
+    where: {
+      isDeleted: true,
+      deleteAt: {
+        lt: beforeDate,
+      },
+    },
+  });
 };
 export default {
   createUser,
@@ -88,4 +93,5 @@ export default {
   restoreUserById,
   softDeleteUserById,
   hardDeleteUserById,
+  cleanUpDeletedUsers,
 };
